requirements.txt
streamlit>=1.33,<2
pandas>=2.2,<3
plotly>=5.20,<6
Pillow>=10,<12
numpy>=1.26,<3
opencv-python>=4.9,<5
pytesseract>=0.3.10,<0.4
pdfplumber>=0.11,<0.12
PyPDF2>=3.0,<4
requests>=2.31,<3
beautifulsoup4>=4.12,<5
